+++ CodeCube.MVC/Attributes/Validation/IntegerValidationAttribute.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.ComponentModel.DataAnnotations;
+
+namespace CodeCube.Mvc.Attributes.Validation
+{
+    [AttributeUsage(AttributeTargets.Property)]
+    public class IntegerValidationAttribute : RegularExpressionAttribute
+    {
+        public IntegerValidationAttribute()
+            : base(@"^\d+$")
+        {
+        }
+    }
+}