+++ CodeCube.MVC/Enumerations/ETextboxType.cs
@@ -1,0 +1,24 @@
+﻿namespace CodeCube.Mvc.Enumerations
+{
+    /// <summary>
+    /// Enumeration for all different types of textboxes available in HTML5.
+    /// </summary>
+    public enum ETextboxType
+    {
+        Text = 0,
+        Email = 1,
+        Url = 2,
+        File = 3,
+        Phone = 4,
+        Search = 5,
+        Number = 6,
+        Range = 7,
+        Date = 8,
+        Month = 9,
+        Week = 10,
+        Time = 11,
+        DateTime = 12,
+        DateTimeLocal = 13,
+        Color = 14
+    }
+}